Enum化追加変更
■config/application.conf
・DBの設定方法をSlickModuleのフォーマットに合わせdbName.profileからslick.dbs.default.profileに変更

■analytics/conf/application.conf
・slick.dbs.default.profileの設定をconfig/application.confを参照するようにしたため削除

■Application.scala
・logIdをIntに変更

■build.sbt
・play-analyticsプロジェクトがmodel-commonとconfigに依存するように変更
・common-propertyがutilに依存するように変更

■ClientLogClassifierFactory.scala
・propertyを取得する処理を共通化
・DestinationTypeを取得処理をMECEに
・windowを更新する処理でrequestBufferを詰める
・Butten event end処理でupdateWithEndを呼び出すときにファイル名、行番号を詰める
・updateWithEndでwindow名を比較する処理を除去

■DefaultDatabaseConfigProvider.scala
・DefaultDBNameとしてplay.slick.db.defaultを参照するのではなくSlickModule.DbKeyConfig+SlickModule.DefaultDbNameを参照するように変更

■EBlockType等のプロパティ
・EnumerationLikeをミックスイン

■EnumerationLik
・withNameOpを追加

■Params
・logIdをIntに変更

+++ jp.co.nri.nefs.tool.analytics/app/controllers/Application.scala
@@ -42,7 +42,7 @@ class Application @Inject() (
     windowDetails.map(wd => Ok(html.list(wd, params)))
   }
 
-  def fileDownload(logId: Long) = Action {
+  def fileDownload(logId: Int) = Action {
     val fileName = Await.result(windowDetailDao.fileName(logId), 10.seconds)
     val fileDir = config.underlying.getString("logDir")
     Ok.sendFile(
